+++ package.json
@@ -1,69 +1,69 @@
 {
-  "name": "snow-shot",
-  "version": "0.1.2",
-  "author": "chao",
-  "private": true,
-  "scripts": {
-    "dev": "next dev --turbopack -p 8301",
-    "build": "next build",
-    "start": "next start",
-    "update:excalidraw": "cd ../excalidraw && yarn build:package && cd ../snow-shot && pnpm update @mg-chao/excalidraw",
-    "lint": "next lint"
-  },
-  "dependencies": {
-    "@ant-design/icons": "^5.6.1",
-    "@ant-design/nextjs-registry": "^1.0.2",
-    "@ant-design/pro-form": "^2.31.7",
-    "@ant-design/v5-patch-for-react-19": "^1.0.3",
-    "@ant-design/x": "^1.2.0",
-    "@mg-chao/excalidraw": "file:../excalidraw/packages/excalidraw",
-    "@tauri-apps/api": "^2.5.0",
-    "@tauri-apps/plugin-autostart": "~2.3.0",
-    "@tauri-apps/plugin-clipboard-manager": "~2",
-    "@tauri-apps/plugin-dialog": "^2.2.1",
-    "@tauri-apps/plugin-fs": "^2.2.1",
-    "@tauri-apps/plugin-global-shortcut": "^2.2.0",
-    "@tauri-apps/plugin-http": "~2.4.3",
-    "@tauri-apps/plugin-log": "^2.4.0",
-    "@tauri-apps/plugin-store": "~2.2.0",
-    "@tweenjs/tween.js": "^25.0.0",
-    "antd": "^5.25.1",
-    "antd-style": "^3.7.1",
-    "color": "^5.0.0",
-    "dayjs": "1.11.13",
-    "es-toolkit": "^1.37.2",
-    "flatbush": "^4.4.0",
-    "js-base64": "^3.7.7",
-    "next": "^15.3.2",
-    "openai": "^4.98.0",
-    "pixi.js": "^8.9.2",
-    "platform": "^1.3.6",
-    "react": "^19.1.0",
-    "react-dom": "^19.1.0",
-    "react-hotkeys-hook": "^4.6.2",
-    "react-intl": "^7.1.11",
-    "react-markdown": "^10.1.0",
-    "react-scrollbars-custom": "^4.1.1",
-    "react-syntax-highlighter": "^15.6.1",
-    "react-truncate-markup": "^5.1.2",
-    "remark-gfm": "^4.0.1"
-  },
-  "devDependencies": {
-    "@eslint/eslintrc": "^3.3.1",
-    "@next/eslint-plugin-next": "^15.3.2",
-    "@tauri-apps/cli": "^2.5.0",
-    "@types/markdown-it": "^14.1.2",
-    "@types/node": "^20.17.46",
-    "@types/react": "^19.1.4",
-    "@types/react-dom": "^19.1.4",
-    "@types/react-syntax-highlighter": "^15.5.13",
-    "eslint": "9.21.0",
-    "eslint-config-next": "15.1.7",
-    "eslint-config-prettier": "^10.1.5",
-    "eslint-plugin-react-hooks": "^5.2.0",
-    "prettier": "^3.5.3",
-    "prettier-plugin-organize-imports": "^4.1.0",
-    "typescript": "^5.8.3"
-  },
-  "packageManager": "pnpm@10.8.1+sha512.c50088ba998c67b8ca8c99df8a5e02fd2ae2e2b29aaf238feaa9e124248d3f48f9fb6db2424949ff901cffbb5e0f0cc1ad6aedb602cd29450751d11c35023677"
+    "name": "snow-shot",
+    "version": "0.1.2",
+    "author": "chao",
+    "private": true,
+    "scripts": {
+        "dev": "next dev --turbopack -p 8301",
+        "build": "next build",
+        "start": "next start",
+        "update:excalidraw": "cd ../excalidraw && yarn build:package && cd ../snow-shot && pnpm update @mg-chao/excalidraw",
+        "lint": "next lint"
+    },
+    "dependencies": {
+        "@ant-design/icons": "^5.6.1",
+        "@ant-design/nextjs-registry": "^1.0.2",
+        "@ant-design/pro-form": "^2.31.7",
+        "@ant-design/v5-patch-for-react-19": "^1.0.3",
+        "@ant-design/x": "^1.3.0",
+        "@mg-chao/excalidraw": "file:../excalidraw/packages/excalidraw",
+        "@tauri-apps/api": "^2.5.0",
+        "@tauri-apps/plugin-autostart": "~2.3.0",
+        "@tauri-apps/plugin-clipboard-manager": "~2.2.2",
+        "@tauri-apps/plugin-dialog": "^2.2.2",
+        "@tauri-apps/plugin-fs": "^2.3.0",
+        "@tauri-apps/plugin-global-shortcut": "^2.2.1",
+        "@tauri-apps/plugin-http": "~2.4.4",
+        "@tauri-apps/plugin-log": "^2.4.0",
+        "@tauri-apps/plugin-store": "~2.2.0",
+        "@tweenjs/tween.js": "^25.0.0",
+        "antd": "^5.25.2",
+        "antd-style": "^3.7.1",
+        "color": "^5.0.0",
+        "dayjs": "1.11.13",
+        "es-toolkit": "^1.38.0",
+        "flatbush": "^4.4.0",
+        "js-base64": "^3.7.7",
+        "next": "^15.3.2",
+        "openai": "^4.100.0",
+        "pixi.js": "^8.9.2",
+        "platform": "^1.3.6",
+        "react": "^19.1.0",
+        "react-dom": "^19.1.0",
+        "react-hotkeys-hook": "^5.1.0",
+        "react-intl": "^7.1.11",
+        "react-markdown": "^10.1.0",
+        "react-scrollbars-custom": "^4.1.1",
+        "react-syntax-highlighter": "^15.6.1",
+        "react-truncate-markup": "^5.1.2",
+        "remark-gfm": "^4.0.1"
+    },
+    "devDependencies": {
+        "@eslint/eslintrc": "^3.3.1",
+        "@next/eslint-plugin-next": "^15.3.2",
+        "@tauri-apps/cli": "^2.5.0",
+        "@types/markdown-it": "^14.1.2",
+        "@types/node": "^20.17.50",
+        "@types/react": "^19.1.5",
+        "@types/react-dom": "^19.1.5",
+        "@types/react-syntax-highlighter": "^15.5.13",
+        "eslint": "9.21.0",
+        "eslint-config-next": "15.1.7",
+        "eslint-config-prettier": "^10.1.5",
+        "eslint-plugin-react-hooks": "^5.2.0",
+        "prettier": "^3.5.3",
+        "prettier-plugin-organize-imports": "^4.1.0",
+        "typescript": "^5.8.3"
+    },
+    "packageManager": "pnpm@10.8.1+sha512.c50088ba998c67b8ca8c99df8a5e02fd2ae2e2b29aaf238feaa9e124248d3f48f9fb6db2424949ff901cffbb5e0f0cc1ad6aedb602cd29450751d11c35023677"
 }
