+++ src/app/tools/chat/components/workflowButton/index.tsx
@@ -8,7 +8,7 @@ import {
 } from '../updateOrCreateWorkflow';
 import { FormattedMessage } from 'react-intl';
 import { useContext, useRef } from 'react';
-import { AntdContext } from '@/app/layout';
+import { AntdContext } from '@/components/globalLayoutExtra';
 import { SendMessageActionType } from '../workflowList';
 
 export const WorkflowButton: React.FC<{
