+++ src/services/tools/index.ts
@@ -5,8 +5,8 @@ import { fetch } from '@tauri-apps/plugin-http';
 export const getUrl = (url: string, params?: Record<string, any>) => {
     let baseUrl;
     if (process.env.NODE_ENV === 'development') {
-        baseUrl = 'http://127.0.0.1:5101/';
-        // baseUrl = 'https://snowshot.mgchao.top/';
+        // baseUrl = 'http://127.0.0.1:5101/';
+        baseUrl = 'https://snowshot.mgchao.top/';
         // baseUrl = 'http://120.79.232.67/';
     } else {
         baseUrl = 'https://snowshot.mgchao.top/';
